backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<2.0.0
python-jose[cryptography]>=3.3.1,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.0.6
python-dotenv>=0.19.0,<0.20.0
alembic>=1.7.4,<2.0.0
email-validator>=1.1.3,<2.0.0
pytest>=6.2.5,<7.0.0
httpx>=0.19.0,<0.20.0
python-dateutil>=2.8.2,<3.0.0
pydantic>=1.8.2,<2.0.0
